% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add30DayInHospitalMortality.R
\name{Add30DayInHospitalMortality}
\alias{Add30DayInHospitalMortality}
\title{Add 30-day in hospital mortality}
\usage{
Add30DayInHospitalMortality(study.sample, from.date = "doar",
  from.time = "toar", to.date = "dodd", to.time = "todd",
  date.format = "\%Y-\%m-\%d", time.format = "\%H:\%M",
  died.variable = "died", died.value = "Yes", levels = c("Yes",
  "No"), variable.name = "m30d", add.as.factor = TRUE,
  drop.used.variables = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{from.date}{Character or POSIXt vector of length 1. The name of the
variable with the date from which to start counting. Defaults to "doar".}

\item{from.time}{Character or POSIXt vector of length 1. The name of the
variable with the time from which to start counting. Defaults to "toar".}

\item{to.date}{Character or POSIXt vector of length 1. The name of the
variable with the date from which to start counting. Defaults to "dodd".}

\item{to.time}{Character or POSIXt vector of length 1. The name of the
variable with the time from which to start counting. Defaults to "todd".}

\item{date.format}{Character vector of length 1. The date format. Defaults to
"\%Y-\%m-\%d".}

\item{time.format}{Character vector of length 1. The date format. Defaults to
"\%H:\%M".}

\item{died.variable}{Character vector of length 1. The name of the variable
indicating whether the patient was dead at the end of follow up. Defaults
to "died".}

\item{died.value}{Character vector of length 1. The value or level of the
died variable that means that the patients died. Defaults to "Yes".}

\item{levels}{Character vector of length 2. The levels to use to encode the
resulting 30-day in hospital mortality variable. The first item in the
vector should be the level to be used to represent a death. Defaults to
c("Yes", "No").}

\item{variable.name}{Character vector of length 1. The name of the 30-day in
hospital mortality variable. Defaults to "m30d".}

\item{add.as.factor}{Logical vector of length 1. If TRUE the 30-day in
hospital mortality variables is added to the study sample as a
factor. If FALSE it is added as character. Defaults to TRUE.}

\item{drop.used.variables}{Logical vector of length 1. If TRUE the date and
time variables used to calculate 30-day in hospital mortality is dropped
from the sample. Defaults to TRUE.}
}
\description{
Adds the variable 30-day in hospital mortality to the study sample
}
