% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSpecificInjuryVariable.R
\name{AddSpecificInjuryVariable}
\alias{AddSpecificInjuryVariable}
\title{Add specific injury variable}
\usage{
AddSpecificInjuryVariable(study.sample, injury.codes, variable.name,
  injury.variables = c("e_1_icd", "e_2_icd", "e_3_icd", "e_4_icd",
  "e_5_icd", "e_6_icd", "e_7_icd", "e_8_icd", "e_9_icd", "e_10_icd",
  "e_11_icd", "e_12_icd", "xray_1_icd", "xray_2_icd", "xray_3_icd",
  "xray_4_icd", "xray_5_icd", "xray_6_icd", "xray_7_icd", "xray_8_icd",
  "xray_9_icd", "xray_10_icd", "xray_11_icd", "fast_1_icd", "fast_2_icd",
  "fast_3_icd", "fast_4_icd", "fast_5_icd", "fast_6_icd", "fast_7_icd",
  "fast_8_icd", "fast_9_icd", "fast_10_icd", "fast_11_icd", "ct_1_icd",
  "ct_2_icd", "ct_3_icd", "ct_4_icd", "ct_5_icd",      "ct_6_icd",
  "ct_7_icd", "ct_8_icd", "ct_9_icd", "ct_10_icd", "ct_11_icd",
  "ct_12_icd", "ct_13_icd", "op_1_icd", "op_2_icd", "op_3_icd", "op_4_icd",
  "op_5_icd", "op_6_icd", "op_7_icd", "op_8_icd", "op_9_icd", "op_10_icd",
  "op_11_icd"), levels = c("Yes", "No"), add.as.factor = TRUE,
  drop.used.variables = FALSE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{injury.codes}{Character vector. The patterns to be used to define the
specific injury or injury pattern. If any of the codes is present in any
of the variables listen in injury.variables an observation will be
classified as having the injury. No default.}

\item{variable.name}{Character vector of length 1. The name of the
variable. No default.}

\item{injury.variables}{Character vector. The names of the variables with
injudy data. Defaults to c("e_1_icd", "e_2_icd", "e_3_icd", "e_4_icd",
"e_5_icd", "e_6_icd", "e_7_icd", "e_8_icd", "e_9_icd", "e_10_icd",
"e_11_icd", "e_12_icd", "xray_1_icd", "xray_2_icd", "xray_3_icd",
"xray_4_icd", "xray_5_icd", "xray_6_icd", "xray_7_icd", "xray_8_icd",
"xray_9_icd", "xray_10_icd", "xray_11_icd", "fast_1_icd", "fast_2_icd",
"fast_3_icd", "fast_4_icd", "fast_5_icd", "fast_6_icd", "fast_7_icd",
"fast_8_icd", "fast_9_icd", "fast_10_icd", "fast_11_icd", "ct_1_icd",
"ct_2_icd", "ct_3_icd", "ct_4_icd", "ct_5_icd", "ct_6_icd", "ct_7_icd",
"ct_8_icd", "ct_9_icd", "ct_10_icd", "ct_11_icd", "ct_12_icd",
"ct_13_icd", "op_1_icd", "op_2_icd", "op_3_icd", "op_4_icd", "op_5_icd",
"op_6_icd", "op_7_icd", "op_8_icd", "op_9_icd", "op_10_icd",
"op_11_icd").}

\item{levels}{Character vector of length 2. The levels to use to encode the
resulting injury variable. The first item in the vector should be the
level to be used to represent having the injury. Defaults to c("Yes",
"No").}

\item{add.as.factor}{Logical vector of length 1. If TRUE the variable is
added to the study sample as a factor. If FALSE it is added as
character. Defaults to TRUE.}

\item{drop.used.variables}{Logical vector of length 1. If TRUE the injury
variables used to calculate the injury variable are dropped from the
sample. Defaults to FALSE.}
}
\description{
Adds a new indicator variable to the study sample, defining what patients had
the specific injury or injury pattern provided.
}
