% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTimeBetweenArrivalAndDischarge.R
\name{AddTimeBetweenArrivalAndDishcarge}
\alias{AddTimeBetweenArrivalAndDishcarge}
\title{Add time between arrival and discharge}
\usage{
AddTimeBetweenArrivalAndDishcarge(study.sample, from.date = "doar",
  from.time = "toar", to.date = "dodd", to.time = "todd",
  date.format = "\%Y-\%m-\%d", time.format = "\%H:\%M",
  units = "days", variable.name = "tadd", drop.used.variables = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{from.date}{Character or POSIXt vector of length 1. The name of the
variable with the date from which to start counting. Defaults to "doar".}

\item{from.time}{Character or POSIXt vector of length 1. The name of the
variable with the time from which to start counting. Defaults to "toar".}

\item{to.date}{Character or POSIXt vector of length 1. The name of the
variable with the date from which to start counting. Defaults to "dodd".}

\item{to.time}{Character or POSIXt vector of length 1. The name of the
variable with the time from which to start counting. Defaults to "todd".}

\item{date.format}{Character vector of length 1. The date format. Defaults to
"\%Y-\%m-\%d".}

\item{time.format}{Character vector of length 1. The date format. Defaults to
"\%H:\%M".}

\item{units}{Character vector of length 1. Should be one of "days", "hours",
"minutes", or "seconds". Defaults to "days".}

\item{variable.name}{Character vector of length 1. The name of the time
between injury and arrival variable. Defaults to "tadd".}

\item{drop.used.variables}{Logical vector of length 1. If TRUE the date and
time variables used to calculate time between injury and arrival are
dropped from the sample. Defaults to TRUE.}
}
\description{
Adds the variable time between arrival to and discharge from participating
centre.
}
