% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTimeBetweenInjuryAndArrival.R
\name{AddTimeBetweenInjuryAndArrival}
\alias{AddTimeBetweenInjuryAndArrival}
\title{Add time between injury and arrival}
\usage{
AddTimeBetweenInjuryAndArrival(study.sample, from.date = "doi",
  from.time = "toi", to.date = "doar", to.time = "toar",
  date.format = "\%Y-\%m-\%d", time.format = "\%H:\%M",
  units = "hours", variable.name = "tbia",
  drop.used.variables = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{from.date}{Character or POSIXt vector of length 1. The name of the
variable with the date from which to start counting. Defaults to "doi".}

\item{from.time}{Character or POSIXt vector of length 1. The name of the
variable with the time from which to start counting. Defaults to "toi".}

\item{to.date}{Character or POSIXt vector of length 1. The name of the
variable with the date from which to start counting. Defaults to "doar".}

\item{to.time}{Character or POSIXt vector of length 1. The name of the
variable with the time from which to start counting. Defaults to "toar".}

\item{date.format}{Character vector of length 1. The date format. Defaults to
"\%Y-\%m-\%d".}

\item{time.format}{Character vector of length 1. The date format. Defaults to
"\%H:\%M".}

\item{units}{Character vector of length 1. Should be one of "days", "hours",
"minutes", or "seconds". Defaults to "hours".}

\item{variable.name}{Character vector of length 1. The name of the time
between injury and arrival variable. Defaults to "tbia".}

\item{drop.used.variables}{Logical vector of length 1. If TRUE the date and
time variables used to calculate time between injury and arrival are
dropped from the sample. Defaults to TRUE.}
}
\description{
Adds the variable time between injury and arrival to participating centre.
}
