% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTriageRevisedTraumaScore.R
\name{AddTriageRevisedTraumaScore}
\alias{AddTriageRevisedTraumaScore}
\title{Add Triage Revised Trauma Score}
\usage{
AddTriageRevisedTraumaScore(study.sample, gcs.name = "gcs_t_1",
  sbp.name = "sbp_1", rr.name = "rr_1", variable.name = "trts",
  drop.used.variables = FALSE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{gcs.name}{Character vector of length 1. The name of the Glasgow coma
scale variable. Defaults to "gcs_t_1".}

\item{sbp.name}{Character vector of length 1. The name of the systolic blood
pressure variable. Defaults to "sbp_1".}

\item{rr.name}{Character vector of length 1. The name of the respiratory rate
variable. Defaults to "rr_1".}

\item{variable.name}{Character vector of length 1. The name of the triage
revised trauma score variable. Defaults to "trts".}

\item{drop.used.variables}{Logical vector of length 1. If TRUE the date and
time variables used to calculate time between injury and arrival are
dropped from the sample. Defaults to FALSE.}
}
\description{
Add the Triage Revised Trauma Score to the sample
}
