% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileResults.R
\name{CompileResults}
\alias{CompileResults}
\title{Compile results}
\usage{
CompileResults(file.format = "docx", delete.results.file = TRUE)
}
\arguments{
\item{file.format}{Character vector of length 1. The file format in which to
save the results. Must be one of "md", "pdf", or "docx". Defaults to
"docx".}

\item{delete.results.file}{Logical vector of length 1. If TRUE the
results.Rds file is deleted once its content has been compiled and
saved. Defaults to TRUE.}
}
\description{
Compiles the results saved to the results.Rds file and saves it to disk in
the format of your choice.
}
