% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSampleCharacteristicsTable.R
\name{create.table.of.sample.characteristics}
\alias{create.table.of.sample.characteristics}
\title{Create sample characteristics table}
\usage{
create.table.of.sample.characteristics(study.sample, data.dictionary = NULL,
  group = NULL, variables = NULL, exclude.variables = NULL,
  include.overall = TRUE, include.missing = TRUE, digits = 1,
  save.to.disk = TRUE, file.format = "docx")
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{data.dictionary}{Not currently used.}

\item{group}{Character vector of length 1. The grouping variable. If NULL the
table is not grouped. Defaults to NULL.}

\item{variables}{Character vector. The names of variables to include in the
table. If NULL all variables in data.dictionary is included. Defaults to
NULL.}

\item{exclude.variables}{Character vector. The names of variables to exclude
from the table. If NULL no variables are excluded. Defaults to NULL.}

\item{include.overall}{Logical vector of length 1. If TRUE an overall column
is included in the tables. Used only if group is not NULL. Defaults to
TRUE.}

\item{include.missing}{Not currently used. Logical vector of length 1. If
TRUE a column with the number (%) of missing values in each variable is
included. Defaults to TRUE.}

\item{digits}{Numeric vector of length 1 greater than or equal to 0. Number
of digits to use when rounding table entries. Defaults to 1.}

\item{save.to.disk}{Logical vector of length 1. If TRUE the table object is
saved to disk. Defaults to TRUE.}

\item{file.format}{Character vector of length 1. The format in which to save
the table to disk. Has to be one of c("pdf", "rmd", "docx"). Defaults to
"docx".}
}
\description{
Creates the sample characteristics table. Wrapper of TableOne.
}
