% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFlowchart.R
\name{CreateFlowchart}
\alias{CreateFlowchart}
\title{CreateFlowchart}
\usage{
CreateFlowchart(flowchart.elements = NULL,
  flowchart.file.path = "./flowchart.rtex", print.tikz = TRUE,
  save.tikz = TRUE, compile.flowchart = TRUE, prettify = FALSE,
  read.from.results = TRUE)
}
\arguments{
\item{flowchart.elements}{Character vector. Content for nodes. Length must be > 1. Defaults to NULL.}

\item{flowchart.file.path}{String. The flowchart file path. Defaults to "./flowchart.rtex".}

\item{print.tikz}{Logical. If TRUE, TikZ code is printed to console. Defaults to TRUE.}

\item{save.tikz}{Logical. If TRUE, TikZ code for flowchart is saved to disk. Defaults to TRUE.}

\item{compile.flowchart}{Logical. If TRUE, the flowchart is compiled using knitr or pdflatex, depending on the file format. Defaults to TRUE}

\item{prettify}{Logical. If TRUE, shadows are used for flowchart nodes. Requires TikZ package shadows.blur. Defaults to FALSE.}

\item{read.from.results}{Logical. If TRUE, flowchart.elements is read from results.Rds. Defaults to TRUE.}
}
\description{
Generates flowchart using TikZ. May be compiled to pdf using knitr or pdflatex.
}
