% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStudyTemplate.R
\name{CreateStudyTemplate}
\alias{CreateStudyTemplate}
\title{Create study template}
\usage{
CreateStudyTemplate(study.name = "My bengaltiger study",
  authors = "Firstname Lastname",
  description = "This is a bengaltiger study.", file.name = "RunStudy",
  path = ".", create.directory = FALSE,
  functions.to.include = c("ImportStudyData"), save.as.function = TRUE,
  function.name = file.name, open = TRUE)
}
\arguments{
\item{study.name}{Character vector of length 1. The name of the
study. Defaults to "My bengaltiger study".}

\item{authors}{Character vector. The names of the study authors. Defaults to
c("Firstname Lastname").}

\item{description}{Character vector of length 1. Short description of the
study. Defaults to "This is a bengaltiger study.".}

\item{file.name}{Character vector of length 1. The file name. Defaults to
"RunStudy".}

\item{path}{Character vector of length 1. The path where the study template
is saved. Defaults to ".", i.e. the current working directory.}

\item{create.directory}{Logical vector of length 1. If TRUE the directory to
which path is pointing is created using dir.create(path) if it does not
already exist. Defaults to FALSE.}

\item{functions.to.include}{Character vector. The names of the functions to
include in the template. Defaults to c("ImportsStudyData").}

\item{save.as.function}{Logical vector of length 1. If TRUE the study
template is structured as a R function. Defaults to TRUE.}

\item{function.name}{Character vector of length 1. The name of the study
function. Is used only if as.functions = TRUE. Defaults to file.name.}

\item{open}{Logical vector of length 1. If TRUE the study template file is
opened using R's file.edit(). Defaults to TRUE.}
}
\description{
Creates a .R file with a custom name where all study steps are documented and
run.
}
