% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateTraumaticBrainInjuryProportion.R
\name{EstimateTraumaticBrainInjuryProportion}
\alias{EstimateTraumaticBrainInjuryProportion}
\title{Estimate traumatic brain injury proportion}
\usage{
EstimateTraumaticBrainInjuryProportion(study.sample, variable.name = "tbi",
  tbi.level = "Yes", digits = 3, bootstrap.confidence.interval = TRUE,
  random.seed = NULL, number.of.bootstrap.samples = 1000,
  save.to.disk = TRUE, return.result = FALSE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{variable.name}{Character vector of length 1. The name of the traumatic
brain injury variable. Defaults to "tbi".}

\item{tbi.level}{Character vector of length 1. The level of the traumatic
brain injury variable that indicates a traumatic brain injury. Defaults
to "Yes".}

\item{digits}{Numeric vector of length 1. Must be a positive integer. The
number of digits to use when rounding the proportion, and if applicable,
the lower and upper bounds of the confidence interval. Defaults to 3.}

\item{bootstrap.confidence.interval}{Logical vector of length 1. If TRUE a
confidence interval is estimated using an emperical bootstrap. Deafults
to TRUE.}

\item{random.seed}{Numeric vector of length 1. Has to be an integer. The seed
to use for random number generation. Only used if
bootstrap.conficence.interval is TRUE. Defaults to NULL.}

\item{number.of.bootstrap.samples}{Numeric vector of length 1. Has to be a
positive integer. The number of bootstrap samples to use. Only used it
bootstrap.confidence.interval is TRUE. Defaults to 1000.}

\item{save.to.disk}{Logical vector of length 1. If TRUE the result is saved
to disk in "results.rmd". Defaults to TRUE.}

\item{return.result}{Logical vector of length 1. If TRUE the result is
returned to the parent environment. Default to FALSE.}
}
\description{
Estimates the proportion of patients with traumatic brain injury.
}
