% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFlowchart.R
\name{FormatText}
\alias{FormatText}
\title{FormatText}
\usage{
FormatText(node.text, read.from.results = TRUE)
}
\arguments{
\item{node.text}{Character vector of length 1. The text for a node in the flowchart. No default.}

\item{read.from.results}{Logical vector of length 1. If TRUE the node.text is an element in flowchart.list in the results object. Defaults to TRUE.}
}
\description{
Read flowchart elements, insert math inline, and escape underscores. If read.from.results special formatting is conducted where newlines are removed and commas are corrected.
}
