% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRevisedTraumaScoreComponents.R
\name{GetRevisedTraumaScoreComponents}
\alias{GetRevisedTraumaScoreComponents}
\title{Get revised trauma score components}
\usage{
GetRevisedTraumaScoreComponents(study.sample, gcs.name = "gcs_t_1",
  sbp.name = "sbp_1", rr.name = "rr_1", return.labels = FALSE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{gcs.name}{Character vector of length 1. The name of the Glasgow coma
scale variable. Defaults to "gcs_t_1".}

\item{sbp.name}{Character vector of length 1. The name of the systolic blood
pressure variable. Defaults to "sbp_1".}

\item{rr.name}{Character vector of length 1. The name of the respiratory rate
variable. Defaults to "rr_1".}

\item{return.labels}{Logical vector of length 1. If TRUE the labels rather
than the values are returned. Defaults to FALSE.}
}
\description{
Gets revised trauma score components
}
