% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitDataset.R
\name{GetSampleIndex}
\alias{GetSampleIndex}
\title{GetSampleIndex}
\usage{
GetSampleIndex(n.obs)
}
\arguments{
\item{n.obs}{A named numeric vector. The number of observations that should
be included in each sample. No default.}
}
\description{
Internal function to get an index of observations to include in each sample
}
