% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICDVariables.R
\name{ICDVariables}
\alias{ICDVariables}
\title{ICD variables}
\usage{
ICDVariables(subset = "all")
}
\arguments{
\item{subset}{Character vector. Must be one or more of "all", "external",
"xray", "fast", "ct", "intraoperative". Defaults "all", which is equal
to c("external", "xray", "fast", "ct", "intraoperative").}
}
\description{
Returns the names of ICD 10 variables in the TITCO dataset
}
\examples{
ICDVariables()
ICDVariables("external")
ICDVariables(c("xray", "ct"))

}
