% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTitcoMySQL.R
\name{ImportTitcoMySQL}
\alias{ImportTitcoMySQL}
\title{Import titco study data from mysql server}
\usage{
ImportTitcoMySQL(mysql.server.name = "127.0.0.1", mysql.server.port = 3307,
  mysql.database = "TITCO", mysql.username = "titco",
  mysql.password = "mangrovetitco", mysql.titco.table = "titco")
}
\arguments{
\item{mysql.server.name}{Character vector of length 1. The hostname or IP of the mysql-server, defaults to 127.0.0.1.}

\item{mysql.server.port}{Integer. The port of mysql-server, defaults is 3307.}

\item{mysql.database}{Character vector of length 1. The name of the database, defaults to TITCO}

\item{mysql.username}{Character vector of length 1. The username for the database, defaults to titco.}

\item{mysql.password}{Character vector of length 1. Password for db access, defaults to mangrovetitco}

\item{mysql.titco.table}{Character vector of length 1. The mysql table name, defaults to titco. To use the smaller sample data set, set this to titco_sample}
}
\description{
Imports the titco study data from a mysql server, if no arguments are set, defaults to mangrove SSH tunnel-configuration.
}
