% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTitcoStudyData.R
\name{ImportTitcoStudyData}
\alias{ImportTitcoStudyData}
\title{Import study data from mysql. To be updated with global variables from configurationfile.}
\usage{
ImportTitcoStudyData(mysql.server.name = "127.0.0.1",
  mysql.server.port = 3307, mysql.database = "TITCO",
  mysql.username = "titco", mysql.password = "mangrovetitco",
  mysql.titco.table = "titco")
}
\arguments{
\item{mysql.server.name}{Character vector of length 1. The hostname or IP of the mysqlserver, for localuse use localhost.}

\item{mysql.server.port}{Integer. The port of mysql-server, default is 3306.}

\item{mysql.database}{Character vector of length 1. The name of the database, TITCO for titco access}

\item{mysql.username}{Character vector of length 1. The username for the database, use titco or ntdb.}

\item{mysql.password}{Character vector of length 1. Password for db access.}
}
\description{
Imports the study data from a mysql server
}
