% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsLength1.R
\name{IsLength1}
\alias{IsLength1}
\title{Is length 1}
\usage{
IsLength1(x)
}
\arguments{
\item{x}{A vector. The object to be checked. No default.}
}
\description{
Checks if a given object is a non-list vector of length 1
}
