% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeRoadTrafficInjuryCategories.R
\name{MergeRoadTrafficInjuryCategories}
\alias{MergeRoadTrafficInjuryCategories}
\title{Merge road traffic injury categories}
\usage{
MergeRoadTrafficInjuryCategories(study.sample, variable.name = "moi")
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{variable.name}{Character vector of length 1. The name of the mechanism
of injury variable. Defaults to "moi".}
}
\description{
Merges road traffic injury categories into one
}
