% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlyBluntTraumaPatients.R
\name{OnlyBluntTraumaPatients}
\alias{OnlyBluntTraumaPatients}
\title{Only patients with blunt trauma}
\usage{
OnlyBluntTraumaPatients(study.sample, injury.type.variable.name = "ti",
  blunt.value = "Blunt", remove.missing = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{injury.type.variable.name}{Character vector of length 1. The name of
the age variable. Defaults to "ti".}

\item{blunt.value}{Character or numeric vector of length 1. The value of the
injury type variable that indicates that a patients had blunt trauma.
Defaults to "Blunt".}

\item{remove.missing}{Logical vector of length 1. If TRUE all observations
with missing injury type, as detected by is.na, are removed from the
sample. Defaults to TRUE.}
}
\description{
Keeps only patients with blunt trauma in the sample.
}
