% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlyEmergencyAdmissions.R
\name{OnlyEmergencyAdmissions}
\alias{OnlyEmergencyAdmissions}
\title{Only emergency admissions}
\usage{
OnlyEmergencyAdmissions(study.sample,
  admission.type.variable.name = "ad_type",
  emergency.value = "Emergency", remove.missing = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{admission.type.variable.name}{Character vector of length 1. The name of
the admission type variable. Defaults to "ad_type".}

\item{emergency.value}{Character or numeric vector of length 1. The value of
the admission type variable that indicates that an admission is an
emergency admission. Defaults to "Emergency".}

\item{remove.missing}{Logical vector of length 1. If TRUE all observations
with missing admission type, as detected by is.na, are removed from the
sample. Defaults to TRUE.}
}
\description{
Keeps only emergency admissions in the sample
}
