% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlyHemothoraxPatients.R
\name{OnlyHemothoraxPatients}
\alias{OnlyHemothoraxPatients}
\title{Only patients with hemothorax}
\usage{
OnlyHemothoraxPatients(study.sample, hemothorax.value = "s27.1",
  remove.missing = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{hemothorax.value}{Character or numeric vector of length 1. The patterns
to be used to define that a patient had a hemothorax. Defaults to
"s27.1".}

\item{remove.missing}{Logical vector of length 1. If TRUE all observations
with data on hemothorax diagnosis, as detected by is.na, are removed from
the sample. Defaults to TRUE.}
}
\description{
Keeps only patients with hemothorax.
}
