% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlyPatientsPresentingAfter.R
\name{OnlyPatientsPresentingAfter}
\alias{OnlyPatientsPresentingAfter}
\title{Only patients presenting later than a specific delay}
\usage{
OnlyPatientsPresentingAfter(study.sample, delay.variable.name = "delay",
  delay.cutoff = 6, remove.missing = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{delay.variable.name}{Character vector of length 1. The name of the
delay variable. Defaults to "delay".}

\item{delay.cutoff}{Numeric vector of length 1. The delay cutoff in
hours. Defaults to 6, i.e. only observations with an delay equal to or
more than 6 hours are kept in the sample.}

\item{remove.missing}{Logical vector of length 1. If TRUE all observations
with missing delay, as detected by is.na, are removed from the
sample. Defaults to TRUE.}
}
\description{
Keeps only the patients whom presented later than a specific delay, where
delay refers to the time between injury and arrival to hospital.
}
