% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlyPatientsPresentingBefore.R
\name{OnlyPatientsPresentingBefore}
\alias{OnlyPatientsPresentingBefore}
\title{Only patients presenting within a specific delay}
\usage{
OnlyPatientsPresentingBefore(study.sample, delay.variable.name = "delay",
  delay.cutoff = 6, remove.missing = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{delay.variable.name}{Character vector of length 1. The name of the
delay variable. Defaults to "delay".}

\item{delay.cutoff}{Numeric vector of length 1. The delay cutoff in
hours. Defaults to 6, i.e. only observations with an delay less than 6
hours are kept in the sample.}

\item{remove.missing}{Logical vector of length 1. If TRUE all observations
with missing delay, as detected by is.na, are removed from the
sample. Defaults to TRUE.}
}
\description{
Keeps only the patients whom presented within a specific delay, where delay
refers to the time between injury and arrival to hospital.
}
