% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlyPolytraumaPatients.R
\name{OnlyPolytraumaPatients}
\alias{OnlyPolytraumaPatients}
\title{Only polytrauma patients}
\usage{
OnlyPolytraumaPatients(sample, ais.variables = c("head_and_neck", "face",
  "chest", "extremities", "external"), ais.cutoff = 2,
  remove.missing = FALSE)
}
\arguments{
\item{sample}{Data frame. The study sample. No default.}

\item{ais.variables}{Character vector. The names of the variables with AIS
scores for each body region. Defaults to c("head_and_neck", "face",
"chest", "extremities", "external").}

\item{ais.cutoff}{Integer vector of length 1 between 1 and 6. The cutoff
above which an injury is counted for polytrauma. Defaults to 2.}

\item{remove.missing}{Logical vector of length 1. If TRUE all observations
with missing AIS, as detected by is.na, are removed from the
sample. Defaults to FALSE.}
}
\description{
Keeps only the polytrauma patients. Polytrauma is here defined as at least
two injuries scored above a certain abbreviated injury scale (AIS) score
cutoff in at least two different body regions.
}
