% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlyTransferredPatients.R
\name{OnlyTransferredPatients}
\alias{OnlyTransferredPatients}
\title{Only transferred patients}
\usage{
OnlyTransferredPatients(study.sample, transfer.variable.name = "tran",
  transfer.value = "Yes", remove.missing = TRUE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{transfer.variable.name}{Character vector of length 1. The name of the
transfer variable. Defaults to "tran".}

\item{transfer.value}{Character or numeric vector of length 1. The value of
the transfer variable that indicates that a patients was
transferred. Defaults to "Yes".}

\item{remove.missing}{Logical vector of length 1. If TRUE all observations
with missing transfer, as detected by is.na, are removed from the
sample. Defaults to TRUE.}
}
\description{
Keeps only patients who were transferred from another health facility in the
sample.
}
