% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveToResults.R
\name{SaveToResults}
\alias{SaveToResults}
\title{Save to results}
\usage{
SaveToResults(output.object, object.name, overwrite = TRUE)
}
\arguments{
\item{output.object}{Character vector of length 1. The output to be saved. No
default.}

\item{object.name}{Character vector of length 1. The name of the output
object in the results object. No default.}

\item{overwrite}{Logical vector of length 1. If TRUE any entry in the results
object with the same name as object.name is overwritten. Defaults to
TRUE.}
}
\description{
Saves the desired output to a results file in the current workind
directory. Output saved in this file can later be compiled as a document.
}
