% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceAdditionalFunctions.R
\name{SourceAdditionalFunctions}
\alias{SourceAdditionalFunctions}
\title{Source additional functions}
\usage{
SourceAdditionalFunctions(path = "./misc/R/")
}
\arguments{
\item{path}{Character vector of length 1. The path to the directory that
holds the additional functions. Defaults to "./misc/R/".}
}
\description{
Sources additional functions for use within the project. These functions may
be user written and intended only for this specific projects, or they may be
functions that are yet to be added to the package.
}
