% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitDataset.R
\name{SplitDataset}
\alias{SplitDataset}
\title{SplitDataset}
\usage{
SplitDataset(study.sample, events = NULL, event.variable.name = NULL,
  event.level = NULL, split.proportions = NULL,
  temporal.split = NULL, remove.missing = FALSE, random.seed = NULL,
  sample.names = NULL, return.data.frame = FALSE)
}
\arguments{
\item{study.sample}{A data.frame. The study sample. No default.}

\item{events}{A numeric vector of the same length as the number of splits or
NULL. Each item should indicate the number of events to be included in
the resulting sample. If NULL split.proportions is used instead. Defaults
to NULL.}

\item{event.variable.name}{A character vector of length 1 or NULL. The name
of the variable that defines an event. Defaults to NULL.}

\item{event.level}{A character or numeric vector of length 1 or NULL. The
level of event.variable that defines an event. Default to NULL.}

\item{split.proportions}{A numeric vector of the same length as the number of
splits or NULL. Each item should indicate the proportion of the dataset
that should be included in the resulting sample. If NULL events is used
instead. Defaults to NULL.}

\item{temporal.split}{Not yet implemented.}

\item{remove.missing}{A logical vector of length 1. If TRUE observations with
missing event data, as detected by is.na(), are removed from the sample
and a warning is issued. If FALSE execution stops if there is missing
event data.  Defaults to FALSE.}

\item{random.seed}{A numeric vector of length 1 or NULL. The random seed to
be used when creating splits. Remember to set the seed outside this
function if you are running other tasks that perform random
operations. Defaults to NULL.}

\item{sample.names}{A character vector of the same length as events or
split.proportions, depending on which is used, or NULL. If NULL the
samples will be called "training" and "test" if two samples are to
created, or "training", "validation", and "test" if three samples are to
be created. Defaults to NULL.}

\item{return.data.frame}{A A logical vector of length 1. If TRUE a single
data.frame is returned. This data.frame includes a new column called
.sample which indicates what sample observations belong to. Defaults to
FALSE.}
}
\value{
A named list with three data frames or a single data frame with an
    added column indicating what sample observations belong to.
}
\description{
Splits the dataset into training, validation, and test samples, although they
can be given different names.
}
