% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Logistic regression}
\usage{
LogisticRegression(study.sample, outcome.name, covariate.names,
  stop.if.missing = TRUE, create.table = TRUE,
  table.options = list(odds.ratio = TRUE, confidence.interval = 0.95,
  p.value = FALSE))
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{outcome.name}{Character vector of length 1. The name of the outcome
variable. No default.}

\item{covariate.names}{Character vector. The names of the covariates. No
default.}

\item{stop.if.missing}{Logical of length 1. If TRUE the execution stops if
there is missing data in the outcome or covariates. Defaults to TRUE.}

\item{create.table}{Logical of length 1. If TRUE a table with the regression
results is created and returned. The apperance of this table can be
modified using table.options. If FALSE the model object is
returned. Defaults to TRUE.}

\item{table.options}{List. Can only include "odds.ratio",
"confidence.interval", and "p.value". Valid values of "odds.ratio" are
TRUE or FALSE. If FALSE the coefficients are included in the table
instead of the odds ratios. Defaults to TRUE. "confidence.interval" has
to be a value between 0 and 1 and governs the width of the confidence
interval. Defaults to 0.95, i.e. a 95% confidence interval is
presented. Valid values of "p.value" are TRUE or FALSE. If TRUE the
p-value is included in the table. Defaults to FALSE.}
}
\description{
Runs a logistic regression.
}
