% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Logistic regression}
\usage{
LogisticRegression(study.sample, outcome.name, covariate.names,
  stop.if.missing = TRUE, create.table = TRUE,
  table.options = list(odds.ratio = TRUE, confidence.interval = 0.95,
  include.p.value = FALSE, digits = 2, save.table = TRUE, table.name =
  NULL, verbose = FALSE), verbose = FALSE)
}
\arguments{
\item{study.sample}{Data frame. The study sample. No default.}

\item{outcome.name}{Character vector of length 1. The name of the outcome
variable. No default.}

\item{covariate.names}{Character vector. The names of the covariates. No
default.}

\item{stop.if.missing}{Logical vector of length 1. If TRUE the execution
stops if there is missing data in the outcome or covariates. Defaults to
TRUE.}

\item{create.table}{Logical vector of length 1. If TRUE a table with the
regression results is created and returned. The apperance of this table
can be modified using table.options. If FALSE the model object is
returned. Defaults to TRUE.}

\item{table.options}{List. Can only include "odds.ratio",
"confidence.interval", "include.intercept", "include.p.value", "digits",
"save.table", "table.name", and "verbose". See
?CreateLogisticRegressionTable for details.}

\item{verbose}{Logical vector of length 1. If TRUE progress is printed as the
function runs. Useful for debugging. Defaults to FALSE.}
}
\description{
Runs a logistic regression.
}
